backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
elasticsearch==8.10.0
pandas==2.1.3
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
numpy==1.24.3
requests==2.31.0
